Rework RMPI directory structure (#50)
