+++ src/RMPI/mpidata.jl
@@ -1,0 +1,76 @@
+"""
+    MPIData(data; kwargs...)
+
+Wrapper used for signaling that this data is part of a distributed
+data structure and communication should happen with MPI.
+
+Keyword arguments:
+* `setup = mpi_one_sided` - controls the communication stratgy
+  * [`mpi_one_sided`](@ref) uses one-sided communication with remote memory access (RMA), sets [`MPIOneSided`](@ref) strategy.
+  * [`mpi_point_to_point`](@ref) uses [`MPIPointTOPoint`](@ref) strategy.
+  * [`mpi_no_exchange`](@ref) sets [`MPINoWalkerExchange`](@ref) strategy. Experimental. Use with caution!
+* `comm = mpi_comm()`
+* `root = mpi_root`
+"""
+struct MPIData{D,S}
+    data::D # local data, e.g. a DVec
+    comm::MPI.Comm
+    root::Int32 # rank of root process
+    isroot::Bool # true if running on root process
+    s::S # type (struct) with further details needed for communication
+
+    function MPIData(data::D, comm, root, s::S) where {D, S<:DistributeStrategy}
+        return new{D,S}(data, comm, root, s.id == root, s)
+    end
+end
+# convenient constructor with setup function
+function MPIData(data; setup=mpi_one_sided, comm=mpi_comm(), root=mpi_root)
+    return setup(data, comm, root)
+end
+
+Base.valtype(md::MPIData{D,S}) where {D,S} = valtype(D)
+Rimu.localpart(md::MPIData) = md.data
+Rimu.StochasticStyle(d::MPIData) = Rimu.StochasticStyle(d.data)
+
+"""
+    length(md::MPIData)
+
+Compute the length of the distributed data on every MPI rank with
+`MPI.Allreduce`. MPI syncronizing.
+"""
+Base.length(md::MPIData) = MPI.Allreduce(length(md.data), +, md.comm)
+
+"""
+    norm(md::MPIData, p=2)
+
+Compute the norm of the distributed data on every MPI rank with `MPI.Allreduce`.
+MPI syncronizing.
+"""
+function LinearAlgebra.norm(md::MPIData, p::Real=2)
+    if p === 2
+        return sqrt(MPI.Allreduce(Rimu.DictVectors.norm_sqr(md.data), +, md.comm))
+    elseif p === 1
+        return MPI.Allreduce(Rimu.DictVectors.norm1(md.data), +, md.comm)
+    elseif p === Inf
+        return MPI.Allreduce(Rimu.DictVectors.normInf(md.data), max, md.comm)
+    else
+        @error "$p-norm of MPIData is not implemented."
+    end
+end
+
+"""
+    walkernumber(md::MPIData)
+
+Compute the walkernumber of the distributed data on every MPI rank with `MPI.Allreduce`.
+MPI syncronizing.
+"""
+function Rimu.walkernumber(md::MPIData)
+    return MPI.Allreduce(walkernumber(md.data), +, md.comm)
+end
+
+function LinearAlgebra.dot(x, md::MPIData)
+    return MPI.Allreduce(x⋅localpart(md), +, md.comm)
+end
+function LinearAlgebra.dot(x, lop, md::MPIData)
+    return MPI.Allreduce(dot(x, lop, localpart(md)), +, md.comm)
+end
